feat: 재시도 횟수를 감소시키는 기능 구현
