+++ README.md
@@ -47,7 +47,7 @@
 재시도
 
 - [x] 재시도 가능 여부를 확인한다.
-- [ ] 재시도 횟수를 감소시킨다.
+- [x] 재시도 횟수를 감소시킨다.
 
 입력
