+++ src/main/java/ladder/domain/Retry.java
@@ -1,0 +1,20 @@
+package ladder.domain;
+
+public class Retry {
+
+    private static final int LIMIT = 0;
+
+    private int value;
+
+    public Retry(final int value) {
+        this.value = value;
+    }
+
+    public void decrease() {
+        value--;
+    }
+
+    public boolean isPossible() {
+        return value > LIMIT;
+    }
+}
